requirements.txt
Flask
line-bot-sdk
opencv-python
imgurpython
pyzbar
Pillow